refactor: RegisterSchoolModal 폴더 및 파일 분리
