+++ src/pages/ClassManagement/LunchMenu/RegisterSchoolModal/PagiNationButton/type.ts
@@ -1,0 +1,11 @@
+import { SearchSchoolList } from '@Types/classManagement/school';
+
+export type PagiNationButtonProps = {
+  hasPage: boolean;
+  schoolList: SearchSchoolList;
+  searchSchool: (page: number) => void;
+};
+
+export type LayoutProps = {
+  hasPage: boolean;
+};
