+++ src/pages/ClassManagement/LunchMenu/RegisterSchoolModal/SearchResult/type.ts
@@ -1,0 +1,6 @@
+import { SearchSchoolList } from '@Types/classManagement/school';
+
+export type SearchResultProps = {
+  isLoading: boolean;
+  schoolList: SearchSchoolList;
+};
